+++ include/docker/client.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "docker_request_creator.hpp"
-#include "docker_response_processor.hpp"
 #include "http.hpp"
+#include "request_creator.hpp"
 #include "request_params.hpp"
+#include "response_processor.hpp"
 #include "returns.hpp"
 
 namespace docker {
