+++ include/docker/response_processor.hpp
