+++ src/query_creator.cpp
@@ -1,4 +1,4 @@
-#include "query_creator.hpp"
+#include "docker/query_creator.hpp"
 
 std::string const kVersion = "/version";
 std::string const kContainers = "/containers";
